Constructed abstract base
